Move Signal into signals package
