+++ core-entities/src/commonMain/kotlin/org/pointyware/artes/entities/signals/derivatives/Image.kt
@@ -1,6 +1,6 @@
 package org.pointyware.artes.entities.signals.derivatives
 
-import org.pointyware.artes.entities.Signal
+import org.pointyware.artes.entities.signals.Signal
 
 /**
  * A collection of points spatially arranged on a plane that usually represent a projection into a
