+++ core-entities/src/commonMain/kotlin/org/pointyware/artes/entities/signals/derivatives/Text.kt
@@ -1,6 +1,6 @@
 package org.pointyware.artes.entities.signals.derivatives
 
-import org.pointyware.artes.entities.Signal
+import org.pointyware.artes.entities.signals.Signal
 
 /**
  * A sequence of symbols used to communicate concepts (whatever the hell those are).
